refactor: 주문 결제 요청 코드 위치 수정
